run lightweight tests

+++ test/1.synthetic.sh
@@ -23,50 +23,55 @@ source "${DIR}/_zfs.file.sh"
 ### Step 2. Configure and launch the Database Lab Engine
 
 ## Prepare database data directory.
-sudo docker rm dblab_pg_initdb || true
+export SYNTHETIC_DATA_DIR="${DLE_TEST_MOUNT_DIR}/${POSTGRES_VERSION}"
 
-sudo docker run \
-  --name dblab_pg_initdb \
-  --label dblab_sync \
-  --label dblab_test \
-  --env PGDATA=/var/lib/postgresql/pgdata \
-  --env POSTGRES_HOST_AUTH_METHOD=trust \
-  --volume ${DLE_TEST_MOUNT_DIR}/${DLE_TEST_POOL_NAME}/data:/var/lib/postgresql/pgdata \
-  --detach \
-  postgres:"${POSTGRES_VERSION}"-alpine
+if [[ ! -d ${SYNTHETIC_DATA_DIR} ]]; then
+  sudo docker rm dblab_pg_initdb || true
 
-check_database_readiness(){
-  sudo docker exec dblab_pg_initdb psql -U postgres -c 'select' > /dev/null 2>&1
-  return $?
-}
+  sudo docker run \
+    --name dblab_pg_initdb \
+    --label dblab_sync \
+    --label dblab_test \
+    --env PGDATA=/var/lib/postgresql/pgdata \
+    --env POSTGRES_HOST_AUTH_METHOD=trust \
+    --volume ${SYNTHETIC_DATA_DIR}/${DLE_TEST_POOL_NAME}/data:/var/lib/postgresql/pgdata \
+    --detach \
+    postgres:"${POSTGRES_VERSION}"-alpine
 
-for i in {1..300}; do
-  check_database_readiness && break || echo "test database is not ready yet"
-  sleep 1
-done
+  check_database_readiness(){
+    sudo docker exec dblab_pg_initdb psql -U postgres -c 'select' > /dev/null 2>&1
+    return $?
+  }
 
-check_database_readiness || (echo "test database is not ready" && exit 1)
+  for i in {1..300}; do
+    check_database_readiness && break || echo "test database is not ready yet"
+    sleep 1
+  done
 
-# Restart container explicitly after initdb to make sure that the server will not receive a shutdown request and queries will not be interrupted.
-sudo docker restart dblab_pg_initdb
+  check_database_readiness || (echo "test database is not ready" && exit 1)
 
-for i in {1..300}; do
-  check_database_readiness && break || echo "test database is not ready yet"
-  sleep 1
-done
+  # Restart container explicitly after initdb to make sure that the server will not receive a shutdown request and queries will not be interrupted.
+  sudo docker restart dblab_pg_initdb
 
-check_database_readiness || (echo "test database is not ready" && exit 1)
+  for i in {1..300}; do
+    check_database_readiness && break || echo "test database is not ready yet"
+    sleep 1
+  done
 
-# Create the test database
-sudo docker exec dblab_pg_initdb psql -U postgres -c 'create database test'
+  check_database_readiness || (echo "test database is not ready" && exit 1)
 
-# Generate data in the test database using pgbench
-# 1,000,000 accounts, ~0.14 GiB of data.
-sudo docker exec dblab_pg_initdb pgbench -U postgres -i -s 10 test
+  # Create the test database
+  sudo docker exec dblab_pg_initdb psql -U postgres -c 'create database test'
 
-# Stop and remove the container
-sudo docker stop dblab_pg_initdb
-sudo docker rm dblab_pg_initdb
+  # Generate data in the test database using pgbench
+  # 1,000,000 accounts, ~0.14 GiB of data.
+  sudo docker exec dblab_pg_initdb pgbench -U postgres -i -s 10 test
+
+  # Stop and remove the container
+  sudo docker stop dblab_pg_initdb
+  sudo docker rm dblab_pg_initdb
+
+fi
 
 configDir="$HOME/.dblab/engine/configs"
 metaDir="$HOME/.dblab/engine/meta"
@@ -85,7 +90,7 @@ yq eval -i '
   .server.port = env(DLE_SERVER_PORT) |
   .provision.portPool.from = env(DLE_PORT_POOL_FROM) |
   .provision.portPool.to = env(DLE_PORT_POOL_TO) |
-  .poolManager.mountDir = env(DLE_TEST_MOUNT_DIR) |
+  .poolManager.mountDir = env(SYNTHETIC_DATA_DIR) |
   del(.retrieval.jobs[] | select(. == "logicalDump")) |
   del(.retrieval.jobs[] | select(. == "logicalRestore")) |
   .databaseContainer.dockerImage = "postgresai/extended-postgres:" + strenv(POSTGRES_VERSION)
@@ -104,8 +109,8 @@ sudo docker run \
   --privileged \
   --publish ${DLE_SERVER_PORT}:${DLE_SERVER_PORT} \
   --volume /var/run/docker.sock:/var/run/docker.sock \
-  --volume ${DLE_TEST_MOUNT_DIR}/${DLE_TEST_POOL_NAME}/dump:${DLE_TEST_MOUNT_DIR}/${DLE_TEST_POOL_NAME}/dump \
-  --volume ${DLE_TEST_MOUNT_DIR}:${DLE_TEST_MOUNT_DIR}/:rshared \
+#  --volume ${SYNTHETIC_DATA_DIR}/${DLE_TEST_POOL_NAME}/dump:${SYNTHETIC_DATA_DIR}/${DLE_TEST_POOL_NAME}/dump \
+  --volume ${SYNTHETIC_DATA_DIR}:${SYNTHETIC_DATA_DIR}/:rshared \
   --volume "${configDir}":/home/dblab/configs:ro \
   --volume "${metaDir}":/home/dblab/meta \
   --volume /sys/kernel/debug:/sys/kernel/debug:rw \
